fix: retours MR, azimuth for grid
